requirements.txt
Flask-Migrate==1.8.0
Flask-SQLAlchemy==2.1
Flask-Script==2.0.5
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
SQLAlchemy==1.0.12
Werkzeug==0.11.5
alembic==0.8.6
itsdangerous==0.24
psycopg2cffi==2.7.2
raven[flask]==5.12.0
whitenoise==3.0
user-agents==1.0.1
PyYAML==3.11
